Fix bugprone-narrowing-conversions
error: narrowing conversion from 'size_t' (aka 'unsigned long') to signed type 'int' is implementation-defined

+++ src/SW_datastructs.c
@@ -902,7 +902,7 @@ void SW_DATA_create_tree(
         yCoords,
         xCoords,
         0,
-        numPoints - 1,
+        (int) numPoints - 1,
         ySize,
         xSize,
         0,
